Moved timer functions exposed for tests to script_timer_private.h
